Filter

+++ Backend/src/main/java/com/example/LocationRepository.java
@@ -2,6 +2,6 @@ package com.example;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface LocationRepository extends PagingAndSortingRepository<Location, Integer> {
+public interface LocationRepository extends PagingAndSortingRepository<Location, Long> {
 
 }
